Stroyt 38 be order controller (#21)
* feature/fix: adding order controller/service/repository/filtering/sorting, fixing material sorting

* feature/refactor: adding mapping for orderMaterialMap, changing order service functionality, adding validations for the materials

* changes/refactor: adding a validation for the order creating, adding orderMaterialMap repository, updating the order creating service, refactoring

* changes/refactor: adding an order service facade, changing order services, adding an enum attribute, refactoring

* changes: adding functionality of user id assigning to orders to newly registered users

* refactor: refactoring constructors, adding 1 more user role, removing a build problem

* changes/refactor/fix: updating order service, removing problems with not getting orders, removing build problems, changing db context model, updating contracts, refactoring

+++ src/Stroytorg.Domain/Specifications/MaterialSpecification.cs
@@ -13,31 +13,31 @@ public class MaterialSpecification : BaseSpecification, ISpecification<Material>
 
     public int? CategoryId { get; set; }
 
-    public double? MinPrice { get; set; }
+    public decimal? MinPrice { get; set; }
 
-    public double? MaxPrice { get; set; }
+    public decimal? MaxPrice { get; set; }
 
-    public double? MinStockAmount { get; set; }
+    public decimal? MinStockAmount { get; set; }
 
-    public double? MaxStockAmount { get; set; }
+    public decimal? MaxStockAmount { get; set; }
 
     public bool? IsFavorite { get; set; }
 
-    public double? MinHeight { get; set; }
+    public decimal? MinHeight { get; set; }
 
-    public double? MaxHeight { get; set; }
+    public decimal? MaxHeight { get; set; }
 
-    public double? MinWidth { get; set; }
+    public decimal? MinWidth { get; set; }
 
-    public double? MaxWidth { get; set; }
+    public decimal? MaxWidth { get; set; }
 
-    public double? MinLength { get; set; }
+    public decimal? MinLength { get; set; }
 
-    public double? MaxLength { get; set; }
+    public decimal? MaxLength { get; set; }
 
-    public double? MinWeight { get; set; }
+    public decimal? MinWeight { get; set; }
 
-    public double? MaxWeight { get; set; }
+    public decimal? MaxWeight { get; set; }
 
     public Expression<Func<Material, bool>> SatisfiedBy()
     {
